[FLINK-3775] [shell] Load Flink configuration before forwarding it
This commit makes sure that the GlobalConfiguration is loaded before the FlinkShell
is started.

Add configDir option to FlinkShell

This allows to configure a configuration directory for the FlinkShell. If the
CLI option is not set, then the system tries to find the configuration directory
using first the FLINK_CONF_DIR environment variable and then the standard directories.

Add Apache license header to dummy flink-conf.yaml file

This closes #1914.