[FLINK-3981] don't log duplicate TaskManager registrations
Duplicate TaskManager registrations shouldn't be logged with Exceptions
in the ResourceManager. Duplicate registrations can happen if the
TaskManager sends out registration messages too fast when the actual
reply is not lost but still in transit.

The ResourceManager should simply acknowledge the duplicate
registrations, leaving it up to the JobManager to decide how to treat
the duplicate registrations (currently it will send an AlreadyRegistered
to the TaskManager).

This closes #2045