[FLINK-2733] [tests] Harden ZooKeeperLeaderElectionTest
Hardens ZooKeeperElectionTest by allowing the testing listener to return
out-dated leader information. This can happen if the ZooKeeper connection
was suspended and the new leader information has not been sent to the
testing listener. In this case, the testing listener will be queried again
to return the actual leader information.

Add debug statements to ZooKeeperLeaderElectionTest.testZooKeeperReelection

This closes #2103.